boot: disable overcurrent protection on Buck3 for CCMP13-DVK board
The Buck3 powers the Wi-Fi chip and internal 3V3 circuitry on the SOM.
On the DVK it is also used for powering external 3V3 circuitry including
the Ethernet PHY and CAN transceiver. This is not recommended due to the
limited power Buck3 offers. Under certain circumstances, especially at
temperatures below -20C, the consumption could trigger the shutdown due
to overcurrent, so disable this protection on the DVK.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/CCMP1-24