ccmp25: set PLL7 as source for SAI4 clock and divisors for audio
Set appropriate divisors, multipliers and frac to get
a frequency of 1671.168 MHz on PLL7.
Use that as source of FLEXGEN25 with divisors 1 and 34
to obtain a frequency of 49.152 MHz on SAI4.
This is an exact multiple of 4 of 12.288 MHz for the master
clock for the MAX98089 codec present on the DVK.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9020