stm32_huk: use a different define to use OTP or test HUK
The default source code bases on a conf.mk defined constant to
determine whether to enter the chunk of test key or the OTP HUK.
The define however cannot be disabled (not defined) from Yocto
so we'd better change the logic and enter the test HUK code if
a new constant CFG_OTP_HUK is not defined.

This way, we can inject the definition of the new constant from
Yocto, if the OTP HUK is wanted instead.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9121