ccmp25: configure eMMC poder domain as always-on
An intermittent issue has been identified when booting from USB, where the eMMC
is not properly detected when reading the U-Boot environment. In the
initialization process, a power cycle is performed as part of the mmc
initialization process, however sometimes the eMMC becomes unresponsive and you
get a timeout error, so to avoid this race condition, this commit sets the
vdd_emmc power domain to always-on to avoid removing power during
initialization.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>