ccmp25: rif: configure PZ2 GPIO as non-secure
The PZ2 GPIO is connected to the TAMP_BUTTON line on the CCMP25-DVK.
While it must be configured as secure if used for tamper detection, this commit
sets it as non-secure by default to allow control from Linux, U-Boot, or other
non-secure software components. This default configuration ensures broader
usability unless tamper functionality is explicitly required.

https://onedigi.atlassian.net/browse/DEL-9572

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>