plat-stm32mp1: re-enable secure services to support cryptographic peripheral
This commit re-enables the necessary secure services required for the correct
operation of a TA or PTA that makes use of the hardware cryptographic
peripheral for encryption and decryption tasks.

These services ensure that access to the cryptographic engine is properly
managed within the secure world, restoring full functionality for secure crypto
operations.

https://onedigi.atlassian.net/browse/DEL-9634

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>